#2137 implemented search, refactored modal to new file
